faulty

+++ calorieCalc.h
@@ -1,7 +1,7 @@
 #ifndef CALORIECALC_H
 #define CALORIECALC_H
 // declare functions
-int DailyCalReq();
+double DailyCalReq();
 
 #endif
